submission_package/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
# Note: PyMuPDF (fitz) removed due to build issues on deployment
# Using PyPDF2 for PDF processing instead

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# LLM Integration
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Utilities
numpy==1.26.2
pandas==2.1.4
aiofiles==23.2.1
